+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # 2.3.1 / Unreleased
   * [Improved] #294 node.load() should resolve 'ok', if node is already loaded
   * [Improved] #293 minExpandLevel does not auto-expand
+  * [Fixed] crash in scrollIntoView() when parent is `window`
 
 # 2.3.0 / 2014-08-17
   * [CHANGED] renamed (undocumented) event 'loaderror' to 'loadError'